+++ lib/src/save_location.dart
@@ -6,6 +6,12 @@ enum SaveLocation {
   /// Save to Documents folder
   documents,
 
+  /// Save to Music folder
+  music,
+
+  /// Save to Videos folder
+  videos,
+
   /// Save to app's private documents directory
   appDocuments,
 }
@@ -17,6 +23,10 @@ extension SaveLocationExtension on SaveLocation {
         return 'Downloads';
       case SaveLocation.documents:
         return 'Documents';
+      case SaveLocation.music:
+        return 'Music';
+      case SaveLocation.videos:
+        return 'Videos';
       case SaveLocation.appDocuments:
         return 'App Documents';
     }
@@ -28,6 +38,10 @@ extension SaveLocationExtension on SaveLocation {
         return 'Saves to the public Downloads folder';
       case SaveLocation.documents:
         return 'Saves to the public Documents folder';
+      case SaveLocation.music:
+        return 'Saves to the public Music folder';
+      case SaveLocation.videos:
+        return 'Saves to the public Videos folder';
       case SaveLocation.appDocuments:
         return 'Saves to the app\'s private documents directory';
     }